backend/requirements.txt
fastapi
uvicorn
python-dotenv
youtube-transcript-api
google-generativeai
pydantic

